refactor(handlers): extract cds-hints.ts from intent.ts (Stage B)
Move CDS impact bucketing/ordering, where-used scoping, update/delete/
activation dependency hints, and the reserved-keyword guard (437 lines,
3 clusters) into one module shared by write/activate/context handlers.
warnCdsReservedKeywords re-exported for back-compat. 3672 tests, A1 zero-diff.
